feat: send email base on the SSO successfully
