--- backend/apps/web/internal/send_mail/config.cfg
@@ -1,4 +1,0 @@
-[azure]
-clientId = 449ba78d-4b78-4e8e-8be1-324f5d0efc72
-tenantId = common
-graphUserScopes = User.Read Mail.Read Mail.Send
