--- backend/apps/web/internal/send_mail/graph.py
@@ -1,138 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-# <UserAuthConfigSnippet>
-from configparser import SectionProxy
-from azure.identity import DeviceCodeCredential
-from msgraph import GraphServiceClient
-from msgraph.generated.users.item.user_item_request_builder import UserItemRequestBuilder
-from msgraph.generated.users.item.mail_folders.item.messages.messages_request_builder import (
-    MessagesRequestBuilder)
-from msgraph.generated.users.item.send_mail.send_mail_post_request_body import (
-    SendMailPostRequestBody)
-from msgraph.generated.models.message import Message
-from msgraph.generated.models.item_body import ItemBody
-from msgraph.generated.models.body_type import BodyType
-from msgraph.generated.models.recipient import Recipient
-from msgraph.generated.models.email_address import EmailAddress
-
-import base64,os
-from msgraph.generated.models.attachment import Attachment
-from msgraph.generated.models.file_attachment import FileAttachment
-
-class Graph:
-    settings: SectionProxy
-    device_code_credential: DeviceCodeCredential
-    user_client: GraphServiceClient
-
-    def __init__(self, config: SectionProxy):
-        self.settings = config
-        client_id = self.settings['clientId']
-        tenant_id = self.settings['tenantId']
-        graph_scopes = self.settings['graphUserScopes'].split(' ')
-
-        self.device_code_credential = DeviceCodeCredential(client_id, tenant_id = tenant_id)
-        self.user_client = GraphServiceClient(self.device_code_credential, graph_scopes)
-# </UserAuthConfigSnippet>
-
-    # <GetUserTokenSnippet>
-    async def get_user_token(self):
-        graph_scopes = self.settings['graphUserScopes']
-        access_token = self.device_code_credential.get_token(graph_scopes)
-        return access_token.token
-    # </GetUserTokenSnippet>
-
-    # <GetUserSnippet>
-    async def get_user(self):
-        # Only request specific properties using $select
-        query_params = UserItemRequestBuilder.UserItemRequestBuilderGetQueryParameters(
-            select=['displayName', 'mail', 'userPrincipalName']
-        )
-
-        request_config = UserItemRequestBuilder.UserItemRequestBuilderGetRequestConfiguration(
-            query_parameters=query_params
-        )
-
-        user = await self.user_client.me.get(request_configuration=request_config)
-        return user
-    # </GetUserSnippet>
-
-    # <GetInboxSnippet>
-    async def get_inbox(self):
-        query_params = MessagesRequestBuilder.MessagesRequestBuilderGetQueryParameters(
-            # Only request specific properties
-            select=['from', 'isRead', 'receivedDateTime', 'subject'],
-            # Get at most 25 results
-            top=25,
-            # Sort by received time, newest first
-            orderby=['receivedDateTime DESC']
-        )
-        request_config = MessagesRequestBuilder.MessagesRequestBuilderGetRequestConfiguration(
-            query_parameters= query_params
-        )
-
-        messages = await self.user_client.me.mail_folders.by_mail_folder_id('inbox').messages.get(
-                request_configuration=request_config)
-        return messages
-    # </GetInboxSnippet>
-
-    # <SendMailSnippet>
-    async def send_mail(self, subject: str, body: str, recipient: str):
-        message = Message()
-        message.subject = subject
-
-        message.body = ItemBody()
-        message.body.content_type = BodyType.Text
-        message.body.content = body
-
-        to_recipient = Recipient()
-        to_recipient.email_address = EmailAddress()
-        to_recipient.email_address.address = recipient
-        message.to_recipients = []
-        message.to_recipients.append(to_recipient)
-
-        request_body = SendMailPostRequestBody()
-        request_body.message = message
-
-        await self.user_client.me.send_mail.post(body=request_body)
-    # </SendMailSnippet>
-
-    #<sendMailSnippet>
-    async def send_leave_mail(self,subject:str,leave_body:str,recipient: str, attachment_path: str,attachment_name: str):
-        message = Message()
-        message.subject = subject
-
-        message.body = ItemBody()
-        message.body.content_type = BodyType.Text
-        message.body.content = leave_body
-
-        to_recipient = Recipient()
-        to_recipient.email_address = EmailAddress()
-        to_recipient.email_address.address = recipient
-        message.to_recipients = []
-        message.to_recipients.append(to_recipient)
-
-        if attachment_path and os.path.isfile(attachment_path):
-            with open(attachment_path, 'rb') as attachment_file:
-                    attachment_content = attachment_file.read()
-                    attachment_base64 = base64.urlsafe_b64encode(attachment_content).decode('utf-8')
-                    attachment = FileAttachment(
-                        odata_type = "#microsoft.graph.fileAttachment",
-                        name=attachment_name,
-                        content_bytes=base64.urlsafe_b64decode(attachment_base64),
-                        content_type="text/plain"
-                    )
-            message.attachments = []
-            message.attachments.append(attachment)
-
-        request_body = SendMailPostRequestBody()
-        request_body.message = message
-        request_body.save_to_sent_items = True
-        await self.user_client.me.send_mail.post(body=request_body)
-    #<SendMailSnippet>        
-
-    # <MakeGraphCallSnippet>
-    async def make_graph_call(self):
-        # INSERT YOUR CODE HERE
-        return
-    # </MakeGraphCallSnippet>
